pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "flatscraper"
version = "0.1.0"
description = "Multi-platform flat search automation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.40.0",
    "groq>=0.4.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
run = "run:main"
